+++ app/src/main/java/wendu/jsbdemo/CallJavascriptActivity.java
@@ -9,37 +9,39 @@ import org.json.JSONObject;
 
 import wendu.dsbridge.DWebView;
 import wendu.dsbridge.OnReturnValue;
+import wendu.dsbridge.WebViewEventImpl;
 
-public class CallJavascriptActivity extends AppCompatActivity implements View.OnClickListener {
+public class CallJavascriptActivity extends AppCompatActivity implements View.OnClickListener{
 
     DWebView dWebView;
+    public WebViewEventImpl mWebViewEvent;
 
-    public <T extends View> T getView(int viewId) {
-        View view = findViewById(viewId);
-        return (T) view;
-    }
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_call_javascript);
-        getView(R.id.addValue).setOnClickListener(this);
-        getView(R.id.append).setOnClickListener(this);
-        getView(R.id.startTimer).setOnClickListener(this);
-        getView(R.id.synAddValue).setOnClickListener(this);
-        getView(R.id.synGetInfo).setOnClickListener(this);
-        getView(R.id.asynAddValue).setOnClickListener(this);
-        getView(R.id.asynGetInfo).setOnClickListener(this);
-        getView(R.id.hasMethodAddValue).setOnClickListener(this);
-        getView(R.id.hasMethodXX).setOnClickListener(this);
-        getView(R.id.hasMethodAsynAddValue).setOnClickListener(this);
-        getView(R.id.hasMethodAsynXX).setOnClickListener(this);
-        DWebView.setWebContentsDebuggingEnabled(true);
-        dWebView= getView(R.id.webview);
-        dWebView.loadUrl("file:///android_asset/native-call-js.html");
+        findViewById(R.id.addValue).setOnClickListener(this);
+        findViewById(R.id.append).setOnClickListener(this);
+        findViewById(R.id.startTimer).setOnClickListener(this);
+        findViewById(R.id.synAddValue).setOnClickListener(this);
+        findViewById(R.id.synGetInfo).setOnClickListener(this);
+        findViewById(R.id.asynAddValue).setOnClickListener(this);
+        findViewById(R.id.asynGetInfo).setOnClickListener(this);
+        findViewById(R.id.hasMethodAddValue).setOnClickListener(this);
+        findViewById(R.id.hasMethodXX).setOnClickListener(this);
+        findViewById(R.id.hasMethodAsynAddValue).setOnClickListener(this);
+        findViewById(R.id.hasMethodAsynXX).setOnClickListener(this);
+
+        dWebView = findViewById(R.id.webview);
+        mWebViewEvent = new WebViewEventImpl(dWebView);
+        mWebViewEvent.setWebContentsDebuggingEnabled(true);
+        mWebViewEvent.loadUrl("file:///android_asset/native-call-js.html");
 
 
     }
+
 
     void showToast(Object o) {
         Toast.makeText(this, o.toString(), Toast.LENGTH_SHORT).show();
@@ -49,7 +51,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
     public void onClick(View v) {
         switch (v.getId()) {
             case R.id.addValue:
-                dWebView.callHandler("addValue", new Object[]{3, 4}, new OnReturnValue<Integer>() {
+                mWebViewEvent.callHandler("addValue", new Object[]{3, 4}, new OnReturnValue<Integer>() {
                     @Override
                     public void onValue(Integer retValue) {
                         showToast(retValue);
@@ -57,7 +59,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.append:
-                dWebView.callHandler("append", new Object[]{"I", "love", "you"}, new OnReturnValue<String>() {
+                mWebViewEvent.callHandler("append", new Object[]{"I", "love", "you"}, new OnReturnValue<String>() {
                     @Override
                     public void onValue(String retValue) {
                         showToast(retValue);
@@ -65,7 +67,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.startTimer:
-                dWebView.callHandler("startTimer", new OnReturnValue<Integer>() {
+                mWebViewEvent.callHandler("startTimer", new OnReturnValue<Integer>() {
                     @Override
                     public void onValue(Integer retValue) {
                         showToast(retValue);
@@ -73,7 +75,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.synAddValue:
-                dWebView.callHandler("syn.addValue", new Object[]{5, 6}, new OnReturnValue<Integer>() {
+                mWebViewEvent.callHandler("syn.addValue", new Object[]{5, 6}, new OnReturnValue<Integer>() {
                     @Override
                     public void onValue(Integer retValue) {
                         showToast(retValue);
@@ -81,7 +83,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.synGetInfo:
-                dWebView.callHandler("syn.getInfo", new OnReturnValue<JSONObject>() {
+                mWebViewEvent.callHandler("syn.getInfo", new OnReturnValue<JSONObject>() {
                     @Override
                     public void onValue(JSONObject retValue) {
                         showToast(retValue);
@@ -89,7 +91,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.asynAddValue:
-                dWebView.callHandler("asyn.addValue", new Object[]{5, 6}, new OnReturnValue<Integer>() {
+                mWebViewEvent.callHandler("asyn.addValue", new Object[]{5, 6}, new OnReturnValue<Integer>() {
                     @Override
                     public void onValue(Integer retValue) {
                         showToast(retValue);
@@ -97,7 +99,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.asynGetInfo:
-                dWebView.callHandler("asyn.getInfo", new OnReturnValue<JSONObject>() {
+                mWebViewEvent.callHandler("asyn.getInfo", new OnReturnValue<JSONObject>() {
                     @Override
                     public void onValue(JSONObject retValue) {
                         showToast(retValue);
@@ -105,7 +107,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.hasMethodAddValue:
-                dWebView.hasJavascriptMethod("addValue", new OnReturnValue<Boolean>() {
+                mWebViewEvent.hasJavascriptMethod("addValue", new OnReturnValue<Boolean>() {
                     @Override
                     public void onValue(Boolean retValue) {
                         showToast(retValue);
@@ -113,7 +115,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.hasMethodXX:
-                dWebView.hasJavascriptMethod("XX", new OnReturnValue<Boolean>() {
+                mWebViewEvent.hasJavascriptMethod("XX", new OnReturnValue<Boolean>() {
                     @Override
                     public void onValue(Boolean retValue) {
                         showToast(retValue);
@@ -121,7 +123,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.hasMethodAsynAddValue:
-                dWebView.hasJavascriptMethod("asyn.addValue", new OnReturnValue<Boolean>() {
+                mWebViewEvent.hasJavascriptMethod("asyn.addValue", new OnReturnValue<Boolean>() {
                     @Override
                     public void onValue(Boolean retValue) {
                         showToast(retValue);
@@ -129,7 +131,7 @@ public class CallJavascriptActivity extends AppCompatActivity implements View.On
                 });
                 break;
             case R.id.hasMethodAsynXX:
-                dWebView.hasJavascriptMethod("asyn.XX", new OnReturnValue<Boolean>() {
+                mWebViewEvent.hasJavascriptMethod("asyn.XX", new OnReturnValue<Boolean>() {
                     @Override
                     public void onValue(Boolean retValue) {
                         showToast(retValue);
