+++ app/src/main/java/wendu/jsbdemo/MainActivity.java
@@ -3,27 +3,14 @@ package wendu.jsbdemo;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.View;
-
-import com.tencent.smtt.sdk.QbSdk;
 
 public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        QbSdk.initX5Environment(this, new QbSdk.PreInitCallback() {
-            @Override
-            public void onCoreInitFinished() {
-                Log.d("jsbridge","onCoreInitFinished");
-            }
 
-            @Override
-            public void onViewInitFinished(boolean b) {
-                Log.d("jsbridge","onCoreInitFinished");
-            }
-        });
+        super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         findViewById(R.id.callJs).setOnClickListener(new View.OnClickListener() {
             @Override
