+++ dsbridge/src/main/java/wendu/dsbridge/JavascriptCloseWindowListener.java
@@ -1,0 +1,8 @@
+package wendu.dsbridge;
+
+public interface JavascriptCloseWindowListener {
+        /**
+         * @return If true, close the current activity, otherwise, do nothing.
+         */
+        boolean onClose();
+    }
