+++ dsbridge/src/main/java/wendu/dsbridge/OnReturnValue.java
@@ -5,5 +5,5 @@ package wendu.dsbridge;
  */
 
 public interface OnReturnValue<T> {
-    void onValue( T retValue);
+    void onValue(T retValue);
 }
